HDDT download delete before write file.

+++ Winform/HoaDonDienTu/MONET/V6ThuePost/ApiConnect/MONET_WS.cs
@@ -146,18 +146,17 @@ namespace V6ThuePostMonetApi
             }
 
             string path = Path.Combine(savefolder, fileName + ".zip");
-            if (File.Exists(path))
+            try
             {
-                try
-                {
-                    File.Delete(path);
-                }
-                catch
-                {
-                    //
-                }
+                if (File.Exists(path)) File.Delete(path);
             }
-            File.WriteAllBytes(path, objFile.fileToBytes);
+            catch
+            {
+            }
+            finally
+            {
+                if (!File.Exists(path)) File.WriteAllBytes(path, objFile.fileToBytes);
+            }
 
             return path;
         }
@@ -191,18 +190,17 @@ namespace V6ThuePostMonetApi
             }
 
             string path = Path.Combine(savefolder, fileName + ".zip");
-            if (File.Exists(path))
+            try
             {
-                try
-                {
-                    File.Delete(path);
-                }
-                catch
-                {
-                    //
-                }
+                if (File.Exists(path)) File.Delete(path);
             }
-            File.WriteAllBytes(path, objFile.fileToBytes);
+            catch
+            {
+            }
+            finally
+            {
+                if (!File.Exists(path)) File.WriteAllBytes(path, objFile.fileToBytes);
+            }
 
             return path;
         }
@@ -239,21 +237,16 @@ namespace V6ThuePostMonetApi
                 throw new Exception("Download no file!");
             }
             string path = Path.Combine(savefolder, fileName + ".pdf");
-
-            if (File.Exists(path))
+            try
             {
-                try
-                {
-                    File.Delete(path);
-                }
-                catch
-                {
-                    //
-                }
+                if (File.Exists(path)) File.Delete(path);
             }
-            if (!File.Exists(path))
+            catch
             {
-                File.WriteAllBytes(path, objFile.fileToBytes);
+            }
+            finally
+            {
+                if (!File.Exists(path)) File.WriteAllBytes(path, objFile.fileToBytes);
             }
             
             return path;
@@ -291,21 +284,16 @@ namespace V6ThuePostMonetApi
             }
 
             string path = Path.Combine(savefolder, fileName + ".pdf");
-
-            if (File.Exists(path))
+            try
             {
-                try
-                {
-                    File.Delete(path);
-                }
-                catch
-                {
-                    //
-                }
+                if (File.Exists(path)) File.Delete(path);
             }
-            if (!File.Exists(path))
+            catch
             {
-                File.WriteAllBytes(path, objFile.fileToBytes);
+            }
+            finally
+            {
+                if (!File.Exists(path)) File.WriteAllBytes(path, objFile.fileToBytes);
             }
 
             return path;
